+++ test/Models/ErrorServiceSpec.ts
@@ -1,20 +1,15 @@
+import { http, HttpResponse } from "msw";
 import TerriaError from "../../lib/Core/TerriaError";
 import { ErrorServiceProvider } from "../../lib/Models/ErrorServiceProviders/ErrorService";
 import StubErrorServiceProvider from "../../lib/Models/ErrorServiceProviders/StubErrorServiceProvider";
-import Terria, { defaultLoadConfig } from "../../lib/Models/Terria";
+import Terria from "../../lib/Models/Terria";
+import { worker } from "../mocks/browser";
 
 describe("ErrorService", function () {
   let mockErrorServiceProvider: ErrorServiceProvider;
   let terria: Terria;
 
   beforeAll(() => {
-    jasmine.Ajax.stubRequest(/.*/).andError({});
-    jasmine.Ajax.stubRequest(/.*(serverconfig|proxyabledomains).*/).andReturn({
-      responseText: JSON.stringify({ foo: "bar" })
-    });
-    jasmine.Ajax.stubRequest("test-config.json").andReturn({
-      responseText: JSON.stringify({ config: true })
-    });
     mockErrorServiceProvider = {
       init: () => {},
       error: () => {}
@@ -22,6 +17,13 @@ describe("ErrorService", function () {
   });
 
   beforeEach(() => {
+    worker.use(
+      http.get("serverconfig", () => HttpResponse.json({ foo: "bar" })),
+      http.get("*/proxyabledomains", () => HttpResponse.json({ foo: "bar" })),
+      http.get("*/test-config.json", () => HttpResponse.json({ config: true })),
+      http.all("*", () => HttpResponse.error())
+    );
+
     terria = new Terria({
       appBaseHref: "/",
       baseUrl: "./"
@@ -31,7 +33,7 @@ describe("ErrorService", function () {
   it("Initializes an error service, passing in config", async function () {
     const initSpy = spyOn(mockErrorServiceProvider, "init");
     await terria.start({
-      loadConfig: () => defaultLoadConfig("test-config.json"),
+      configUrl: "test-config.json",
       errorService: mockErrorServiceProvider
     });
     expect(initSpy).toHaveBeenCalledTimes(1);
@@ -40,7 +42,7 @@ describe("ErrorService", function () {
   it("Gets called with error", async function () {
     const errorSpy = spyOn(mockErrorServiceProvider, "error").and.callThrough();
     await terria.start({
-      loadConfig: () => defaultLoadConfig("test-config.json"),
+      configUrl: "test-config.json",
       errorService: mockErrorServiceProvider
     });
     const error = new TerriaError({
@@ -50,9 +52,9 @@ describe("ErrorService", function () {
     expect(errorSpy).toHaveBeenCalledWith(error);
   });
 
-  it("Falls back to stub provider", () => {
-    terria.start({
-      loadConfig: () => defaultLoadConfig("test-config.json")
+  it("Falls back to stub provider", async () => {
+    await terria.start({
+      configUrl: "test-config.json"
     });
     expect(terria.errorService).toEqual(jasmine.any(StubErrorServiceProvider));
   });